First round of PR fixes
