+++ src/github/pullRequestManager.ts
@@ -144,6 +144,9 @@ export class PullRequestManager implements vscode.Disposable {
 	private _onDidChangeMentionableUsers = new vscode.EventEmitter<IAccount[]>();
 	readonly onDidChangeMentionableUsers: vscode.Event<IAccount[]> = this._onDidChangeMentionableUsers.event;
 
+	private _onDidChangeAssignableUsers = new vscode.EventEmitter<IAccount[]>();
+	readonly onDidChangeAssignableUsers: vscode.Event<IAccount[]> = this._onDidChangeAssignableUsers.event;
+
 	private _state: PRManagerState = PRManagerState.Initializing;
 
 	constructor(
@@ -480,6 +483,17 @@ export class PullRequestManager implements vscode.Disposable {
 		}
 	}
 
+	getAllAssignableUsers(): IAccount[] | undefined {
+		if (this._assignableUsers) {
+			const allAssignableUsers: IAccount[] = [];
+			Object.keys(this._assignableUsers).forEach(k => {
+				allAssignableUsers.push(...this._assignableUsers![k]);
+			});
+
+			return allAssignableUsers;
+		}
+	}
+
 	async getMentionableUsers(clearCache?: boolean): Promise<{ [key: string]: IAccount[] }> {
 		if (clearCache) {
 			delete this._mentionableUsers;
@@ -491,12 +505,12 @@ export class PullRequestManager implements vscode.Disposable {
 
 		if (!this._fetchMentionableUsersPromise) {
 			const cache: { [key: string]: IAccount[] } = {};
-			const mentionableUsers: IAccount[] = [];
+			const allMentionableUsers: IAccount[] = [];
 			return this._fetchMentionableUsersPromise = new Promise((resolve) => {
 				const promises = this._githubRepositories.map(async githubRepository => {
 					const data = await githubRepository.getMentionableUsers();
 					cache[githubRepository.remote.remoteName] = data;
-					mentionableUsers.push(...data);
+					allMentionableUsers.push(...data);
 					return;
 				});
 
@@ -504,7 +518,7 @@ export class PullRequestManager implements vscode.Disposable {
 					this._mentionableUsers = cache;
 					this._fetchMentionableUsersPromise = undefined;
 					resolve(cache);
-					this._onDidChangeMentionableUsers.fire(mentionableUsers);
+					this._onDidChangeMentionableUsers.fire(allMentionableUsers);
 				});
 			});
 		}
@@ -523,10 +537,12 @@ export class PullRequestManager implements vscode.Disposable {
 
 		if (!this._fetchAssignableUsersPromise) {
 			const cache: { [key: string]: IAccount[] } = {};
+			const allAssignableUsers: IAccount[] = [];
 			return this._fetchAssignableUsersPromise = new Promise((resolve) => {
 				const promises = this._githubRepositories.map(async githubRepository => {
 					const data = await githubRepository.getAssignableUsers();
 					cache[githubRepository.remote.remoteName] = data.sort(loginComparator);
+					allAssignableUsers.push(...data);
 					return;
 				});
 
@@ -534,6 +550,7 @@ export class PullRequestManager implements vscode.Disposable {
 					this._assignableUsers = cache;
 					this._fetchAssignableUsersPromise = undefined;
 					resolve(cache);
+					this._onDidChangeAssignableUsers.fire(allAssignableUsers);
 				});
 			});
 		}
