+++ src/github/queries.gql
@@ -538,6 +538,7 @@ query GetAssignableUsers($owner: String!, $name: String!, $first: Int!, $after: 
 				avatarUrl
 				name
 				url
+				email
 			}
 			pageInfo {
 				hasNextPage
